[IntersectionObserver] If necessary, schedule a frame after delay
If the following sequence happens:

- IntersectionObserver sends a notification
- Layout is dirtied in a way that should generate a new notification
- Lifecyce update runs
  - IntersectionObserver skips processing due to its 'delay' parameter

... then we need to guarantee that the IntersectionObserver will send
the necessary notification after the required delay. If animation
frames are being regularly generated, this won't be a problem, because
the observer will get a chance to run at the end of every lifecycle
update. But if the document goes dormant, and frames are not being
regularly generated, then we must force a lifecycle update after enough
time has passed, to give the observer a chance to run.

Change-Id: I74b63486e5c9b05c662f9e2e6e1c0b4baf00eba5